feat: payment, reservation 분리 - #49
