+++ src/main/java/com/permitseoul/permitserver/domain/coupon/CouponBaseException.java
@@ -1,0 +1,4 @@
+package com.permitseoul.permitserver.domain.coupon;
+
+public abstract class CouponBaseException extends RuntimeException {
+}
