+++ src/main/java/com/permitseoul/permitserver/domain/coupon/core/domain/entity/CouponEntity.java
@@ -23,9 +23,13 @@ public class CouponEntity {
     @Column(name = "discount_rates")
     private int discountRates;
 
+    @Column(name = "is_used")
+    private boolean isUsed;
+
     private CouponEntity(long eventId, String couponCode, int discountRates) {
         this.eventId = eventId;
         this.couponCode = couponCode;
         this.discountRates = discountRates;
+        this.isUsed = false;
     }
 }
