+++ src/main/java/com/permitseoul/permitserver/domain/coupon/core/exception/CouponConflictException.java
@@ -1,0 +1,4 @@
+package com.permitseoul.permitserver.domain.coupon.core.exception;
+
+public class CouponConflictException extends CouponCoreException {
+}
