+++ src/main/java/com/permitseoul/permitserver/domain/coupon/core/exception/CouponCoreException.java
@@ -1,0 +1,6 @@
+package com.permitseoul.permitserver.domain.coupon.core.exception;
+
+import com.permitseoul.permitserver.domain.coupon.CouponBaseException;
+
+public abstract class CouponCoreException extends CouponBaseException {
+}
