+++ src/main/java/com/permitseoul/permitserver/domain/coupon/core/repository/CouponRepository.java
@@ -1,0 +1,13 @@
+package com.permitseoul.permitserver.domain.coupon.core.repository;
+
+import com.permitseoul.permitserver.domain.coupon.core.domain.entity.CouponEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface CouponRepository extends JpaRepository<CouponEntity, Long> {
+    boolean existsByCouponCode(final String couponCode);
+    boolean existsByCouponCodeAndUsedFalse(String couponCode);
+}
