+++ src/main/java/com/permitseoul/permitserver/domain/reservation/api/service/ReservationService.java
@@ -1,146 +1,98 @@
 package com.permitseoul.permitserver.domain.reservation.api.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.permitseoul.permitserver.domain.coupon.core.component.CouponRetriever;
+import com.permitseoul.permitserver.domain.coupon.core.exception.CouponConflictException;
+import com.permitseoul.permitserver.domain.coupon.core.exception.CouponNotfoundException;
 import com.permitseoul.permitserver.domain.event.core.component.EventRetriever;
 import com.permitseoul.permitserver.domain.event.core.domain.Event;
 import com.permitseoul.permitserver.domain.event.core.exception.EventNotfoundException;
 import com.permitseoul.permitserver.domain.payment.api.client.TossPaymentClient;
-import com.permitseoul.permitserver.domain.payment.api.dto.PaymentCancelResponse;
-import com.permitseoul.permitserver.domain.payment.api.dto.TossPaymentRequest;
-import com.permitseoul.permitserver.domain.payment.api.dto.PaymentResponse;
-import com.permitseoul.permitserver.domain.payment.api.dto.TossConfirmErrorResponse;
 import com.permitseoul.permitserver.domain.payment.core.component.PaymentRetriever;
 import com.permitseoul.permitserver.domain.payment.core.component.PaymentSaver;
 import com.permitseoul.permitserver.domain.payment.core.component.PaymentUpdater;
-import com.permitseoul.permitserver.domain.payment.core.domain.Currency;
-import com.permitseoul.permitserver.domain.payment.core.domain.Payment;
-import com.permitseoul.permitserver.domain.payment.core.domain.PaymentStatus;
-import com.permitseoul.permitserver.domain.payment.core.domain.entity.PaymentEntity;
-import com.permitseoul.permitserver.domain.payment.core.exception.PaymentNotFoundException;
 import com.permitseoul.permitserver.domain.paymentcancel.core.component.PaymentCancelSaver;
 import com.permitseoul.permitserver.domain.reservation.api.TossProperties;
-import com.permitseoul.permitserver.domain.reservation.api.dto.TossPaymentCancelRequest;
-import com.permitseoul.permitserver.domain.reservation.api.dto.PaymentConfirmResponse;
 import com.permitseoul.permitserver.domain.reservation.api.dto.PaymentReadyRequest;
-import com.permitseoul.permitserver.domain.reservation.api.dto.PaymentReadyResponse;
-import com.permitseoul.permitserver.domain.reservation.api.exception.*;
+import com.permitseoul.permitserver.domain.reservation.api.dto.ReservationInfoResponse;
+import com.permitseoul.permitserver.domain.reservation.api.exception.ConflictReservationException;
+import com.permitseoul.permitserver.domain.reservation.api.exception.NotfoundReservationException;
 import com.permitseoul.permitserver.domain.reservation.core.component.ReservationRetriever;
 import com.permitseoul.permitserver.domain.reservation.core.component.ReservationSaver;
 import com.permitseoul.permitserver.domain.reservation.core.component.ReservationUpdater;
 import com.permitseoul.permitserver.domain.reservation.core.domain.Reservation;
-import com.permitseoul.permitserver.domain.reservation.core.domain.ReservationStatus;
-import com.permitseoul.permitserver.domain.reservation.core.domain.entity.ReservationEntity;
-import com.permitseoul.permitserver.domain.reservation.core.exception.ReservationNotFoundException;
 import com.permitseoul.permitserver.domain.reservationticket.core.component.ReservationTicketRetriever;
 import com.permitseoul.permitserver.domain.reservationticket.core.component.ReservationTicketSaver;
-import com.permitseoul.permitserver.domain.reservationticket.core.domain.ReservationTicket;
 import com.permitseoul.permitserver.domain.ticket.core.component.TicketRetriever;
 import com.permitseoul.permitserver.domain.ticket.core.component.TicketSaver;
 import com.permitseoul.permitserver.domain.ticket.core.component.TicketUpdater;
-import com.permitseoul.permitserver.domain.ticket.core.domain.Ticket;
-import com.permitseoul.permitserver.domain.ticket.core.domain.TicketStatus;
-import com.permitseoul.permitserver.domain.ticket.core.domain.entity.TicketEntity;
-import com.permitseoul.permitserver.domain.ticket.core.exception.TicketNotFoundException;
 import com.permitseoul.permitserver.domain.tickettype.core.component.TicketTypeRetriever;
-import com.permitseoul.permitserver.domain.tickettype.core.domain.entity.TicketTypeEntity;
-import com.permitseoul.permitserver.domain.tickettype.core.exception.TicketTypeNotfoundException;
-import com.permitseoul.permitserver.domain.tickettype.core.exception.TicketTypeInsufficientCountException;
-import com.permitseoul.permitserver.global.TicketCodeGenerator;
 import com.permitseoul.permitserver.domain.user.core.component.UserRetriever;
 import com.permitseoul.permitserver.domain.user.core.domain.User;
 import com.permitseoul.permitserver.domain.user.core.exception.UserNotFoundException;
-import com.permitseoul.permitserver.global.exception.AlgorithmException;
-import com.permitseoul.permitserver.global.exception.DateFormatException;
-import com.permitseoul.permitserver.global.exception.IllegalEnumTransitionException;
-import com.permitseoul.permitserver.global.formatter.DateFormatterUtil;
 import com.permitseoul.permitserver.global.response.code.ErrorCode;
-import feign.FeignException;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
-import java.util.Base64;
 import java.util.List;
-import java.util.stream.IntStream;
-import java.util.zip.DataFormatException;
-
-import static com.permitseoul.permitserver.global.formatter.DateFormatterUtil.parseDateToLocalDateTime;
 
 @Service
-@EnableConfigurationProperties(TossProperties.class)
+@RequiredArgsConstructor
 public class ReservationService {
-    private static final String COLON = ":";
-    private static final String AUTH_TYPE_BASIC = "Basic ";
-    private static final String JSON_PARSING_FAIL = "JSON 파싱 실패입니다.";
-    private static final String PAYMENT_FEIGN_FAIL = "결제 Feign 통신 오류입니다.";
-    private static final String CANCEL_REASON = "고객 요구 취소";
 
     private final ReservationSaver reservationSaver;
     private final ReservationTicketSaver reservationTicketSaver;
-    private final ReservationTicketRetriever reservationTicketRetriever;
     private final EventRetriever eventRetriever;
     private final UserRetriever userRetriever;
-    private final TossPaymentClient tossPaymentClient;
-    private final ReservationRetriever reservationRetriever;
-    private final TossProperties tossProperties;
-    private final String authorizationHeader;
-    private final PaymentSaver paymentSaver;
-    private final TicketSaver ticketSaver;
     private final TicketTypeRetriever ticketTypeRetriever;
-    private final PaymentRetriever paymentRetriever;
-    private final PaymentCancelSaver paymentCancelSaver;
-    private final PaymentUpdater paymentUpdater;
-    private final TicketUpdater ticketUpdater;
-    private final TicketRetriever ticketRetriever;
-    private final ReservationUpdater reservationUpdater;
-
-
-    public ReservationService(ReservationSaver reservationSaver,
-                              ReservationTicketSaver reservationTicketSaver,
-                              ReservationTicketRetriever reservationTicketRetriever,
-                              EventRetriever eventRetriever,
-                              UserRetriever userRetriever,
-                              TossPaymentClient tossPaymentClient,
-                              ReservationRetriever reservationRetriever,
-                              TossProperties tossProperties,
-                              PaymentSaver paymentSaver,
-                              TicketSaver ticketSaver,
-                              TicketTypeRetriever ticketTypeRetriever,
-                              PaymentRetriever paymentRetriever,
-                              PaymentCancelSaver paymentCancelSaver,
-                              PaymentUpdater paymentUpdater,
-                              TicketUpdater ticketUpdater,
-                              TicketRetriever ticketRetriever,
-                              ReservationUpdater reservationUpdater) {
-        this.reservationSaver = reservationSaver;
-        this.reservationTicketSaver = reservationTicketSaver;
-        this.reservationTicketRetriever = reservationTicketRetriever;
-        this.eventRetriever = eventRetriever;
-        this.userRetriever = userRetriever;
-        this.tossPaymentClient = tossPaymentClient;
-        this.reservationRetriever = reservationRetriever;
-        this.tossProperties = tossProperties;
-        this.authorizationHeader = buildAuthHeader(tossProperties.apiSecretKey());
-        this.paymentSaver = paymentSaver;
-        this.ticketSaver = ticketSaver;
-        this.ticketTypeRetriever = ticketTypeRetriever;
-        this.paymentRetriever = paymentRetriever;
-        this.paymentCancelSaver = paymentCancelSaver;
-        this.paymentUpdater = paymentUpdater;
-        this.ticketUpdater = ticketUpdater;
-        this.ticketRetriever = ticketRetriever;
-        this.reservationUpdater = reservationUpdater;
-    }
+    private final CouponRetriever couponRetriever;
 
     @Transactional
-    public PaymentReadyResponse getPaymentReady(final long userId,
-                                                final long eventId,
-                                                final String couponCode,
-                                                final BigDecimal totalAmount,
-                                                final String orderId,
-                                                final List<PaymentReadyRequest.TicketTypeInfo> ticketTypeInfos) {
+    public String saveReservation(final long userId,
+                                  final long eventId,
+                                  final String couponCode,
+                                  final BigDecimal totalAmount,
+                                  final String orderId,
+                                  final List<PaymentReadyRequest.TicketTypeInfo> ticketTypeInfos) {
+        try {
+            isExistUser(userId);
+            isExistEvent(eventId);
+            isExistTicketType(ticketTypeInfos);
+            if(couponCode != null) {
+                isUsableCouponCode(couponCode);
+            }
+
+            //여기서 터지는 dataintegrationException은 글로벌 핸들러에서 잡고있음.
+            final Reservation reservation = reservationSaver.saveReservation(userId, eventId, orderId, totalAmount, couponCode);
+            ticketTypeInfos.forEach(
+                    ticketTypeInfo -> reservationTicketSaver.saveReservationTicket(ticketTypeInfo.id(), reservation.getOrderId(), ticketTypeInfo.count())
+            );
+            return reservation.getOrderId();
+        } catch (EventNotfoundException e) {
+            throw new NotfoundReservationException(ErrorCode.NOT_FOUND_EVENT);
+        } catch (UserNotFoundException e) {
+            throw new NotfoundReservationException(ErrorCode.NOT_FOUND_USER);
+        } catch (CouponNotfoundException e) {
+            throw new NotfoundReservationException(ErrorCode.NOT_FOUND_COUPON_CODE);
+        } catch (CouponConflictException e) {
+            throw new ConflictReservationException(ErrorCode.CONFLICT_ALREADY_USED_COUPON_CODE);
+        }
+    }
+
+
+
+
+
+
+
+    @Transactional
+    public ReservationInfoResponse getPaymentReady(final long userId,
+                                                   final long eventId,
+                                                   final String couponCode,
+                                                   final BigDecimal totalAmount,
+                                                   final String orderId,
+                                                   final List<PaymentReadyRequest.TicketTypeInfo> ticketTypeInfos) {
         final Event event;
         final User user;
         try {
@@ -152,7 +104,7 @@ public class ReservationService {
             ticketTypeInfos.forEach(
                     ticketTypeInfo -> reservationTicketSaver.saveReservationTicket(ticketTypeInfo.id(), reservation.getOrderId(), ticketTypeInfo.count())
             );
-            return PaymentReadyResponse.of(event.getName(), reservation.getOrderId(), user.getName(), user.getEmail(), reservation.getTotalAmount(), user.getSocialId());
+            return ReservationInfoResponse.of(event.getName(), reservation.getOrderId(), user.getName(), user.getEmail(), reservation.getTotalAmount(), user.getSocialId());
         } catch (EventNotfoundException e) {
             throw new NotfoundReservationException(ErrorCode.NOT_FOUND_EVENT);
         } catch (UserNotFoundException e) {
@@ -160,176 +112,22 @@ public class ReservationService {
         }
     }
 
-    @Transactional
-    public PaymentConfirmResponse getPaymentConfirm(final long userId,
-                                                    final String orderId,
-                                                    final String paymentKey,
-                                                    final BigDecimal totalAmount) {
-        try {
-            final Reservation reservation = reservationRetriever.findReservationByOrderIdAndAmount(orderId, totalAmount, userId);
-            final Event event = eventRetriever.findEventById(reservation.getEventId());
-            final PaymentResponse paymentResponse = getTossPaymentConfirm(authorizationHeader, paymentKey, reservation.getOrderId(), reservation.getTotalAmount());
-
-
-            final Payment savedPayment = savePaymentInfo(reservation, paymentResponse);
-
-            final List<ReservationTicket> findReservationTicketList = reservationTicketRetriever.findAllByOrderId(savedPayment.getOrderId());
-            checkTicketCountAndDecrease(findReservationTicketList);
-            final List<Ticket> newTicketList = generateTickets(findReservationTicketList, userId, reservation);
-            ticketSaver.saveTickets(newTicketList);
-
-            return PaymentConfirmResponse.of(
-                    event.getName(),
-                    DateFormatterUtil.formatEventDate(event.getStartDate(), event.getEndDate())
-            );
-        } catch (ReservationNotFoundException e) {
-            throw new NotfoundReservationException(ErrorCode.NOT_FOUND_RESERVATION);
-        } catch(EventNotfoundException e) {
-            throw new NotfoundReservationException(ErrorCode.NOT_FOUND_EVENT);
-        } catch (TicketTypeNotfoundException e) {
-            throw new NotfoundReservationException(ErrorCode.NOT_FOUND_TICKET_TYPE);
-        } catch(FeignException e) {
-            throw handleFeignException(e);
-        } catch (AlgorithmException e) {
-            throw new TicketAlgorithmException(ErrorCode.INTERNAL_TICKET_ALGORITHM_ERROR);
-        } catch (TicketTypeInsufficientCountException e) {
-            throw new ConflictReservationException(ErrorCode.CONFLICT_INSUFFICIENT_TICKET);
-        } catch (IllegalEnumTransitionException e) {
-            throw new ReservationIllegalException(ErrorCode.INTERNAL_TRANSITION_ENUM_ERROR);
-        }
+    private void isExistUser(final long userId) {
+        userRetriever.isExistUserByUserId(userId);
     }
 
-    @Transactional
-    public void cancelPayment(final long userId, final String orderId) {
-        try {
-            final PaymentEntity paymentEntity = paymentRetriever.findPaymentEntityByOrderId(orderId);
-            final List<TicketEntity> ticketEntity = ticketRetriever.findAllTicketsByOrderId(paymentEntity.getOrderId());
-            final ReservationEntity reservationEntity = reservationRetriever.findReservationEntityById(paymentEntity.getReservationId());
-
-            validateCancelPaymentWithUserId(userId, paymentEntity.getReservationId());
-            cancelTossPaymentAndSave(paymentEntity.getPaymentId(), paymentEntity.getPaymentKey(), paymentEntity.getCurrency());
-
-            updatePaymentStatus(paymentEntity, PaymentStatus.CANCELLED);
-            updateTicketStatus(ticketEntity, TicketStatus.CANCELED);
-            updateReservationStatus(reservationEntity, ReservationStatus.PAYMENT_CANCELED);
-        } catch (PaymentNotFoundException e) {
-            throw new NotfoundReservationException(ErrorCode.NOT_FOUND_PAYMENT);
-        } catch (ReservationNotFoundException e) {
-            throw new NotfoundReservationException(ErrorCode.NOT_FOUND_RESERVATION);
-        } catch (TicketNotFoundException e) {
-            throw new NotfoundReservationException(ErrorCode.NOT_FOUND_TICKET);
-        }
+    private void isExistEvent(final long eventId) {
+        eventRetriever.isExistByEventId(eventId);
     }
 
-    private void updateReservationStatus(final ReservationEntity reservationEntity, final ReservationStatus reservationStatus) {
-        reservationUpdater.updateReservationStatus(reservationEntity, reservationStatus);
-    }
-
-    private void updatePaymentStatus(final PaymentEntity paymentEntity, final PaymentStatus paymentStatus) {
-        paymentUpdater.updatePaymentStatus(paymentEntity ,paymentStatus);
-    }
-
-    private void updateTicketStatus(final List<TicketEntity> ticketEntity, final TicketStatus ticketStatus) {
-        ticketEntity.forEach(
-                ticket -> ticketUpdater.updateTicketStatus(ticket, ticketStatus)
+    private void isExistTicketType(final List<PaymentReadyRequest.TicketTypeInfo> ticketTypeInfos) {
+        ticketTypeInfos.forEach(
+                ticketType -> ticketTypeRetriever.isExistByTicketTypeId(ticketType.id())
         );
     }
 
-    private void cancelTossPaymentAndSave(final long paymentId, final String paymentKey, final Currency currency) {
-        final PaymentCancelResponse cancelResponse = tossPaymentClient.cancelPayment(
-                authorizationHeader,
-                paymentKey,
-                TossPaymentCancelRequest.of(CANCEL_REASON, currency)
-        );
-
-        final PaymentCancelResponse.CancelDetail latestCancelPayment = DateFormatterUtil.getLatestCancelPaymentByDate(cancelResponse.cancels()).orElseThrow(
-                () -> new DateFormatException(ErrorCode.INTERNAL_ISO_DATE_ERROR)
-        );
-        paymentCancelSaver.savePaymentCancel(
-                paymentId,
-                latestCancelPayment.cancelAmount(),
-                latestCancelPayment.transactionKey(),
-                parseDateToLocalDateTime(latestCancelPayment.canceledAt())
-        );
-    }
-
-    private void validateCancelPaymentWithUserId(final long userId, final long reservationId) {
-        final Reservation reservation = reservationRetriever.findReservationById(reservationId);
-        if (reservation.getUserId() != userId) {
-            throw new ReservationUnAuthorizedException(ErrorCode.UNAUTHORIZED_CANCEL_PAYMENT);
-        }
-    }
-
-    private String buildAuthHeader(final String secretKey) {
-        return AUTH_TYPE_BASIC + Base64.getEncoder().encodeToString((secretKey + COLON).getBytes());
-    }
-
-    private PaymentResponse getTossPaymentConfirm(final String authorizationHeader,
-                                                  final String paymentKey,
-                                                  final String orderId,
-                                                  final BigDecimal totalAmount) {
-        return  tossPaymentClient.purchaseConfirm(
-                authorizationHeader,
-                TossPaymentRequest.of(paymentKey, orderId, totalAmount)
-        );
-    }
-
-    private void checkTicketCountAndDecrease(final List<ReservationTicket> reservationTicketList) {
-        reservationTicketList.forEach(reservationTicket -> {
-            final TicketTypeEntity ticketTypeEntity = ticketTypeRetriever.findTicketTypeEntityById(reservationTicket.getTicketTypeId());
-            ticketTypeEntity.decreaseRemainCount(reservationTicket.getCount());
-        });
-    }
-
-    private Payment savePaymentInfo(final Reservation reservation, final PaymentResponse paymentResponse) {
-        return  paymentSaver.savePayment(
-                reservation.getReservationId(),
-                reservation.getOrderId(),
-                reservation.getEventId(),
-                paymentResponse.paymentKey(),
-                reservation.getTotalAmount(),
-                paymentResponse.currency(),
-                DateFormatterUtil.parseDateToLocalDateTime(paymentResponse.requestedAt()),
-                DateFormatterUtil.parseDateToLocalDateTime(paymentResponse.approvedAt())
-        );
-    }
-
-    private List<Ticket> generateTickets(final List<ReservationTicket> reservationTicketList, final long userId, final Reservation reservation) {
-        return reservationTicketList.stream()
-                .flatMap(reservationTicket ->
-                        IntStream.range(0, reservationTicket.getCount())
-                                .mapToObj(i -> Ticket.builder()
-                                        .userId(userId)
-                                        .orderId(reservationTicket.getOrderId())
-                                        .ticketTypeId(reservationTicket.getTicketTypeId())
-                                        .eventId(reservation.getEventId())
-                                        .ticketCode(TicketCodeGenerator.generateTicketCode())
-                                        .isUsed(false)
-                                        .status(TicketStatus.RESERVED)
-                                        .build()
-                                )
-                )
-                .toList();
-    }
-
-    private RuntimeException handleFeignException(final FeignException e) {
-        final String body = e.contentUTF8();
-        if (body != null && !body.isBlank()) {
-            try {
-                final ObjectMapper mapper = new ObjectMapper();
-                final TossConfirmErrorResponse tossError = mapper.readValue(body, TossConfirmErrorResponse.class);
-                return new TossPaymentConfirmException(ErrorCode.INTERNAL_PAYMENT_FEIGN_ERROR, tossError.getMessage());
-            } catch (JsonProcessingException jsonException) {
-                return new PaymentFeignException(
-                        ErrorCode.INTERNAL_JSON_FORMAT_ERROR,
-                        e.getCause() != null ? e.getCause().getMessage() : JSON_PARSING_FAIL
-                );
-            }
-        } else {
-            return new PaymentFeignException(
-                    ErrorCode.INTERNAL_PAYMENT_FEIGN_ERROR,
-                    e.getCause() != null ? e.getCause().getMessage() : PAYMENT_FEIGN_FAIL
-            );
-        }
+    private void isUsableCouponCode(final String couponCode) {
+        couponRetriever.isExistCoupon(couponCode);
+        couponRetriever.isCouponValid(couponCode);
     }
 }
