+++ src/main/java/com/permitseoul/permitserver/domain/reservationticket/core/component/ReservationTicketSaver.java
@@ -11,10 +11,9 @@ import org.springframework.stereotype.Component;
 public class ReservationTicketSaver {
     private final ReservationTicketRepository reservationTicketRepository;
 
-    public ReservationTicket saveReservationTicket(final long ticketTypeId,
+    public void saveReservationTicket(final long ticketTypeId,
                                                    final String orderId,
                                                    final int count) {
-        final ReservationTicketEntity reservationTicketEntity = reservationTicketRepository.save(ReservationTicketEntity.create(ticketTypeId, orderId, count));
-        return ReservationTicket.fromEntity(reservationTicketEntity);
+        reservationTicketRepository.save(ReservationTicketEntity.create(ticketTypeId, orderId, count));
     }
 }
