+++ src/main/java/com/permitseoul/permitserver/domain/tickettype/core/component/TicketTypeRetriever.java
@@ -5,13 +5,22 @@ import com.permitseoul.permitserver.domain.tickettype.core.exception.TicketTypeN
 import com.permitseoul.permitserver.domain.tickettype.core.repository.TicketTypeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 @Component
 @RequiredArgsConstructor
 public class TicketTypeRetriever {
     private final TicketTypeRepository ticketTypeRepository;
 
+    @Transactional(readOnly = true)
     public TicketTypeEntity findTicketTypeEntityById(final long ticketTypeId) {
         return ticketTypeRepository.findById(ticketTypeId).orElseThrow(TicketTypeNotfoundException::new);
+    }
+
+    @Transactional(readOnly = true)
+    public void isExistByTicketTypeId(final long ticketTypeId) {
+        if(!ticketTypeRepository.existsById(ticketTypeId)) {
+            throw new TicketTypeNotfoundException();
+        }
     }
 }
